Move all init functions into separate files
